Refactor NRE loss logic to Strategy Pattern (Phase 2)
